+++ packages/ui/src/graph/index.ts
@@ -1,2 +1,1 @@
 export { Graph } from './graph';
-
