+++ packages/ui/src/graph/models/index.ts
@@ -1,2 +1,1 @@
 export { Renderable } from './renderable';
-
