--- packages/ui/src/renderables/index.ts
@@ -1,2 +1,0 @@
-export { RenderableNode } from './node';
-export { RenderableLink } from './link';
